requirements.txt
aiodns==3.0.0
aiohttp==3.8.3
aiosignal==1.3.1
APScheduler==3.6.3
async-generator==1.10
async-timeout==4.0.2
backports.zoneinfo==0.2.1
cachetools==4.2.2
ccxt==2.4.77
charset-normalizer==2.1.1
cryptography==38.0.4
exceptiongroup==1.1.0
frozenlist==1.3.3
google-auth==2.15.0
google-auth-oauthlib==0.8.0
gspread==5.7.2
h11==0.14.0
iniconfig==1.1.1
ipytest==0.13.0
mistune==0.8.4
multidict==6.0.4
numpy==1.24.1
oauthlib==3.2.2
outcome==1.2.0
pandas==1.5.2
pluggy==1.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==4.3.0
PySocks==1.7.1
pytest==7.2.0
python-telegram-bot==13.15
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
pywin32==305.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.7.2
sortedcontainers==2.4.0
tomli==2.0.1
tornado==6.1
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
webencodings==0.5.1
wincertstore==0.2
wsproto==1.2.0
yarl==1.8.2
